requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
backports.csv==1.0.7
bcrypt==3.1.7
beautifulsoup4==4.8.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.5
django-allauth==0.40.0
django-cors-headers==3.1.0
django-heroku==0.3.1
django-import-export==1.2.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
Fabric3==1.14.post1
gunicorn==19.7.1
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.6.3
paramiko==2.6.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
PyNaCl==1.3.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.4
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
tablib==0.13.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.3.1
xlrd==1.2.0
xlwt==1.3.0
amqp==2.3.2
asn1crypto==0.24.0
astroid==1.6.5
autopep8==1.3.5
backports.functools-lru-cache==1.5
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.4
boto3==1.5.8
botocore==1.8.22
cached-property==1.3.1
celery==4.2.1
certifi==2018.1.18
chardet==3.0.4
cliapp==1.20170827
cmdtest==0.32
configparser==3.5.0
crcmod==1.7
cryptography==2.1.4
Django==1.11.15
django-allauth==0.36.0
django-bulma==0.5.1
django-categories==1.6
django-heroku==0.3.1
django-js-asset==1.1.0
django-mptt==0.9.0
django-object-tools==1.11.0
django-storages==1.6.5
docker==2.5.1
docker-compose==1.17.1
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
enum34==1.1.6
Fabric==1.14.0
Fabric3==1.14.post1
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
isort==4.3.4
jmespath==0.9.3
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kombu==4.2.1
lazy-object-proxy==1.3.1
Markdown==2.6.9
mccabe==0.6.1
mock==2.0.0
oauthlib==2.1.0
paramiko==2.4.1
pbr==3.1.1
pep8==1.7.1
Pillow==6.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.3
pycodestyle==2.4.0
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.26.1
pylint==1.9.2
PyNaCl==1.2.1
pyOpenSSL==17.5.0
python-dateutil==2.7.3
python-decouple==3.1
python-openid==2.2.5
python3-openid==3.1.0
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.11.0
texttable==0.9.1
ttystatus==0.38
unicode-slugify==0.1.3
Unidecode==1.0.22
urllib3==1.23
vine==1.1.4
virtualenv==16.0.0
websocket-client==0.44.0
whitenoise==3.3.1
wrapt==1.10.11
